Started to add affine transform function. Added _CRT_SECURE_NO_WARNINGS Added FIA_DrawImageFromSrcToDst and FIA_DrawImageToDst to the drawing functions. Theses should mimic the win32 BitBlt and .Net Removed warnings. Added C# wrappers for affine transform functions. Added DrawImage Test cases.

+++ src/FreeImageAlgorithms_DistanceTransform.cpp
@@ -203,9 +203,9 @@ FIA_DistanceMap (int width, int height, int normalise)
         for(int x = 0; x < width; x++) {
 
             if (x <= center_x)
-                current_min = x;
+                current_min = (float) x;
             else
-                current_min = width - x;
+                current_min = (float) (width - x);
 
 #ifdef WIN32
             if (y <= center_y)
